+++ app/src/main/java/com/example/letssopt/designsystem/component/Bar/WatchaMainTopBar.kt
@@ -1,4 +1,4 @@
-package com.example.letssopt.designsystem.component
+package com.example.letssopt.designsystem.component.Bar
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -10,6 +10,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import com.example.letssopt.designsystem.component.Button.WatchaTopIconButton
 
 @Composable
 fun WatchaMainTopBar(
