+++ app/src/main/java/com/example/letssopt/designsystem/component/Folder/WatchaFolders.kt
@@ -1,4 +1,4 @@
-package com.example.letssopt.designsystem.component
+package com.example.letssopt.designsystem.component.Folder
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
