+++ app/src/main/java/com/example/letssopt/designsystem/component/Text/WatchaSemiTitle.kt
@@ -1,4 +1,4 @@
-package com.example.letssopt.designsystem.component
+package com.example.letssopt.designsystem.component.Text
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -7,6 +7,7 @@ import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
+import com.example.letssopt.R
 import com.example.letssopt.designsystem.theme.TextPrimary
 
 @Composable
@@ -16,7 +17,7 @@ fun WatchaSemiTitle(
 ){
     Text(
         text = text,
-        fontFamily = FontFamily(Font(com.example.letssopt.R.font.pretendard_bold)),
+        fontFamily = FontFamily(Font(R.font.pretendard_bold)),
         fontWeight = FontWeight.Bold,
         fontSize = 20.sp,
         color = TextPrimary,
