+++ app/src/main/java/com/example/letssopt/presentation/MainActivity.kt
@@ -28,12 +28,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.letssopt.R
-import com.example.letssopt.designsystem.component.WatchaBanner
-import com.example.letssopt.designsystem.component.WatchaBottomBar
-import com.example.letssopt.designsystem.component.WatchaContents
-import com.example.letssopt.designsystem.component.WatchaMainTopBar
-import com.example.letssopt.designsystem.component.WatchaPartyCards
-import com.example.letssopt.designsystem.component.WatchaSemiTitle
+import com.example.letssopt.designsystem.component.Banner.WatchaBanner
+import com.example.letssopt.designsystem.component.Bar.WatchaBottomBar
+import com.example.letssopt.designsystem.component.Content.WatchaContents
+import com.example.letssopt.designsystem.component.Bar.WatchaMainTopBar
+import com.example.letssopt.designsystem.component.Party.WatchaPartyCards
+import com.example.letssopt.designsystem.component.Text.WatchaSemiTitle
 import com.example.letssopt.designsystem.data.bannerImages
 import com.example.letssopt.designsystem.data.contentImages
 import com.example.letssopt.designsystem.data.icons
